pyproject.toml
[tool.poetry]
name = "app"
version = "0.2.0"
description = ""
authors = ["Dennis"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.30.6"
python-multipart = "^0.0.10"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.0"
python-decouple = "^3.8"
bcrypt = "^4.2.1"
pymemcache = "^4.0.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.7.1"
faker = "^30.8.1"
edgedb = "^2.1.0"
pytest-playwright = "^0.6.2"


[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'


[tool.ruff]
line-length = 79
extend-exclude = ['Database', 'migrations', 'queries']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['F401', 'F403', 'F405']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.taskipy.tasks]
run = 'fastapi dev app/main.py'

pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'

lint = 'ruff check .; ruff check . --diff'
check = 'ruff check . --fix; ruff format .'



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
